+++ HouseholdAccountBook/Adapters/Logger/Log.cs
@@ -1,5 +1,5 @@
-﻿using HouseholdAccountBook.Others;
-using HouseholdAccountBook.Properties;
+﻿using HouseholdAccountBook.Properties;
+using HouseholdAccountBook.Utilites;
 using System;
 using System.Collections;
 using System.Diagnostics;
