+++ HouseholdAccountBook/App.xaml.cs
@@ -4,7 +4,7 @@ using HouseholdAccountBook.Adapters.Logger;
 using HouseholdAccountBook.DbHandler;
 using HouseholdAccountBook.Enums;
 using HouseholdAccountBook.Extensions;
-using HouseholdAccountBook.Others;
+using HouseholdAccountBook.Utilites;
 using HouseholdAccountBook.Views.Windows;
 using System;
 using System.Diagnostics;
