+++ HouseholdAccountBook/Args/RequestEventArgs/AddActionListRequestEventArgs.cs
@@ -1,12 +1,13 @@
-﻿using System;
+﻿using HouseholdAccountBook.ViewModels.Component;
+using System;
 using System.Collections.Generic;
 
-namespace HouseholdAccountBook.Others.RequestEventArgs
+namespace HouseholdAccountBook.Args.RequestEventArgs
 {
     /// <summary>
-    /// 移動追加要求イベント時のイベント引数
+    /// 帳簿項目リスト追加要求イベント時のイベント引数
     /// </summary>
-    public class AddMoveRequestEventArgs : DbRequestEventArgsBase
+    public class AddActionListRequestEventArgs : DbRequestEventArgsBase
     {
         /// <summary>
         /// 初期選択する帳簿のID
@@ -20,6 +21,10 @@ namespace HouseholdAccountBook.Others.RequestEventArgs
         /// 初期選択する日付
         /// </summary>
         public DateTime? InitialDate { get; set; }
+        /// <summary>
+        /// 初期表示するCSVレコードリスト
+        /// </summary>
+        public List<CsvViewModel> InitialRecordList { get; set; }
 
         /// <summary>
         /// 登録完了時イベントハンドラ
