+++ HouseholdAccountBook/Args/RequestEventArgs/CopyMoveRequestEventArgs.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace HouseholdAccountBook.Others.RequestEventArgs
+namespace HouseholdAccountBook.Args.RequestEventArgs
 {
     /// <summary>
     /// 移動複製要求イベント時のイベント引数
