+++ HouseholdAccountBook/Args/RequestEventArgs/EditActionRequestEventArgs.cs
@@ -1,15 +1,15 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace HouseholdAccountBook.Others.RequestEventArgs
+namespace HouseholdAccountBook.Args.RequestEventArgs
 {
     /// <summary>
-    /// 帳簿項目複製要求イベント時のイベント引数
+    /// 帳簿項目編集要求イベント時のイベント引数
     /// </summary>
-    public class CopyActionRequestEventArgs : DbRequestEventArgsBase
+    public class EditActionRequestEventArgs : DbRequestEventArgsBase
     {
         /// <summary>
-        /// 複製対象の帳簿項目のID
+        /// 編集対象の帳簿項目のID
         /// </summary>
         public int TargetActionId { get; set; }
