+++ HouseholdAccountBook/Args/RequestEventArgs/SaveFileDialogRequestEventArgs.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace HouseholdAccountBook.Others.RequestEventArgs
+namespace HouseholdAccountBook.Args.RequestEventArgs
 {
     /// <summary>
     /// ファイル保存ダイアログ要求時イベントのイベント引数
