+++ HouseholdAccountBook/Resources/UpdateLog.txt
@@ -2,6 +2,9 @@
 凡例：＋追加 △変更 ×削除 ↑バグ修正
 ---------------------------------------------------------------------------------------------------------------------
 * 2026
+2026-01-14
+  全体
+    △ファイルの配置を変更した - refs #28
 2026-01-13
   全体
     △祝日リストの取得に失敗した場合の通知に表示時間を設定するように変更した - refs #23
