+++ HouseholdAccountBook/ViewModels/Abstract/WindowPartViewModelBase.cs
@@ -1,6 +1,6 @@
 ﻿using HouseholdAccountBook.Adapters.DbHandler;
-using HouseholdAccountBook.Others;
-using HouseholdAccountBook.Others.RequestEventArgs;
+using HouseholdAccountBook.Args.RequestEventArgs;
+using HouseholdAccountBook.Utilites;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
