+++ HouseholdAccountBook/ViewModels/Windows/CsvComparisonWindowViewModel.cs
@@ -5,10 +5,11 @@ using HouseholdAccountBook.Adapters.Dao.DbTable;
 using HouseholdAccountBook.Adapters.DbHandler.Abstract;
 using HouseholdAccountBook.Adapters.Dto.Others;
 using HouseholdAccountBook.Adapters.Logger;
+using HouseholdAccountBook.Args;
+using HouseholdAccountBook.Args.RequestEventArgs;
 using HouseholdAccountBook.Enums;
 using HouseholdAccountBook.Extensions;
-using HouseholdAccountBook.Others;
-using HouseholdAccountBook.Others.RequestEventArgs;
+using HouseholdAccountBook.Utilites;
 using HouseholdAccountBook.ViewModels.Abstract;
 using HouseholdAccountBook.ViewModels.Component;
 using System;
