+++ HouseholdAccountBook/ViewModels/Windows/MainWindowViewModel.cs
@@ -8,11 +8,12 @@ using HouseholdAccountBook.Adapters.Dto.Abstract;
 using HouseholdAccountBook.Adapters.Dto.DbTable;
 using HouseholdAccountBook.Adapters.Dto.KHDbTable;
 using HouseholdAccountBook.Adapters.Logger;
+using HouseholdAccountBook.Args;
+using HouseholdAccountBook.Args.RequestEventArgs;
 using HouseholdAccountBook.DbHandler;
 using HouseholdAccountBook.Enums;
 using HouseholdAccountBook.Extensions;
-using HouseholdAccountBook.Others;
-using HouseholdAccountBook.Others.RequestEventArgs;
+using HouseholdAccountBook.Utilites;
 using HouseholdAccountBook.ViewModels.Abstract;
 using HouseholdAccountBook.ViewModels.Component;
 using HouseholdAccountBook.ViewModels.WindowsParts;
