+++ HouseholdAccountBook/ViewModels/WindowsParts/SettingsWindowItemTabViewModel.cs
@@ -3,8 +3,9 @@ using HouseholdAccountBook.Adapters.Dao.DbTable;
 using HouseholdAccountBook.Adapters.DbHandler.Abstract;
 using HouseholdAccountBook.Adapters.Dto.DbTable;
 using HouseholdAccountBook.Adapters.Logger;
+using HouseholdAccountBook.Args;
 using HouseholdAccountBook.Enums;
-using HouseholdAccountBook.Others;
+using HouseholdAccountBook.Utilites;
 using HouseholdAccountBook.ViewModels.Abstract;
 using HouseholdAccountBook.ViewModels.Component;
 using HouseholdAccountBook.ViewModels.Settings;
