+++ HouseholdAccountBook/Views/Windows/DbSettingWindow.xaml.cs
@@ -1,6 +1,6 @@
 ﻿using HouseholdAccountBook.Adapters.Logger;
 using HouseholdAccountBook.Extensions;
-using HouseholdAccountBook.Others;
+using HouseholdAccountBook.Utilites;
 using System;
 using System.Windows;
 using System.Windows.Controls;
