+++ HouseholdAccountBook/Views/Windows/MainWindow.xaml.cs
@@ -3,7 +3,7 @@ using HouseholdAccountBook.Adapters.DbHandler;
 using HouseholdAccountBook.Adapters.Logger;
 using HouseholdAccountBook.Enums;
 using HouseholdAccountBook.Extensions;
-using HouseholdAccountBook.Others;
+using HouseholdAccountBook.Utilites;
 using HouseholdAccountBook.ViewModels.Component;
 using System;
 using System.ComponentModel;
