+++ HouseholdAccountBook/Views/Windows/TermWindow.xaml.cs
@@ -1,7 +1,7 @@
 ﻿using HouseholdAccountBook.Adapters.DbHandler;
 using HouseholdAccountBook.Adapters.Logger;
 using HouseholdAccountBook.Extensions;
-using HouseholdAccountBook.Others;
+using HouseholdAccountBook.Utilites;
 using System;
 using System.Windows;
 using System.Windows.Controls;
