+++ HouseholdAccountBook/Views/Windows/VersionWindow.xaml.cs
@@ -1,6 +1,6 @@
 ﻿using HouseholdAccountBook.Adapters.Logger;
 using HouseholdAccountBook.Extensions;
-using HouseholdAccountBook.Others;
+using HouseholdAccountBook.Utilites;
 using System.Windows;
 
 namespace HouseholdAccountBook.Views.Windows
